requirements.txt
pywinauto==0.6.8
pytest==7.2.0
allure-pytest==2.13.0
pytest-assume==2.4.3
retrying==1.3.4
requests==2.31.0
